+++ gaios/Recovery/Controllers/PhraseNoteDownViewModel.swift
@@ -1,0 +1,6 @@
+class PhraseNoteDownViewModel {
+    var phrase: String
+    init (_ phrase: String) {
+        self.phrase = phrase
+    }
+}