+++ gaios/Utility/AssetExpandableSelect/Cells/AccountSelectSubCellModel.swift
@@ -1,8 +1,0 @@
-import Foundation
-import UIKit
-import gdk
-
-struct AccountSelectSubCellModel {
-
-    var account: WalletItem
-}