+++ gaios/Utility/SecuritySelect/Cells/OnBoardInfoCellModel.swift
@@ -1,9 +1,0 @@
-import Foundation
-import UIKit
-
-struct OnBoardInfoCellModel {
-
-    var icon: UIImage
-    var title: String
-    var hint: String
-}