+++ gaios/WalletTabBar/ManualBackup/ManualBackupViewController.swift
@@ -2,6 +2,7 @@ import Foundation
 import UIKit
 import LocalAuthentication
 import core
+import greenaddress
 
 class ManualBackupViewController: UIViewController {
 
@@ -11,23 +12,27 @@ class ManualBackupViewController: UIViewController {
     @IBOutlet weak var lblInfo3: UILabel!
     @IBOutlet weak var btnNext: UIButton!
     var showBip85: Bool = false
-    var backupFlowType = BackupFlowType.phrase
+
+    weak var subAccountCreateDelegate: AccountCreateRecoveryKeyDelegate?
+    var viewModel: ManualBackupViewModel!
+
+    var wm: WalletManager { WalletManager.current! }
 
     override func viewDidLoad() {
         super.viewDidLoad()
 
         setContent()
         setStyle()
-        title = "id_back_up_your_wallet".localized
+        title = viewModel.navTitle
         navigationItem.backButtonTitle = "id_back".localized
     }
 
     func setContent() {
-        lblTitle.text = "id_manual_backup".localized
-        lblInfo1.text = "id_offline_written_backup".localized
-        lblInfo2.text = "id_access_wallet_on_a_new_device".localized
-        lblInfo3.text = "id_instant_recovery".localized
-        btnNext.setTitle("Back Up Recovery Phrase", for: .normal)
+        lblTitle.text = viewModel.title
+        lblInfo1.text = viewModel.info1
+        lblInfo2.text = viewModel.info2
+        lblInfo3.text = viewModel.info3
+        btnNext.setTitle(viewModel.btnTitle, for: .normal)
     }
 
     func setStyle() {
@@ -37,16 +42,61 @@ class ManualBackupViewController: UIViewController {
         lblInfo3.setStyle(.txt)
         btnNext.setStyle(.primary)
     }
+
+    func next(_ length: MnemonicLengthOption) {
+        var mnemonic: String?
+        if length == MnemonicLengthOption._24 {
+            mnemonic = try? generateMnemonic()
+        } else {
+            mnemonic = try? generateMnemonic12()
+        }
+        if let mnemonic, let vc = AccountNavigator.phraseNoteDown(phrase: mnemonic,
+        subAccountCreateDelegate: subAccountCreateDelegate) {
+            navigationController?.pushViewController(vc, animated: true)
+        }
+    }
+    func selectLength() {
+        let storyboard = UIStoryboard(name: "Dialogs", bundle: nil)
+        if let vc = storyboard.instantiateViewController(withIdentifier: "DialogListViewController") as? DialogListViewController {
+            vc.delegate = self
+            vc.viewModel = DialogListViewModel(title: viewModel.chooselengthTitle, type: .phrasePrefs, items: PhrasePrefs.getItems())
+            vc.modalPresentationStyle = .overFullScreen
+            present(vc, animated: false, completion: nil)
+        }
+
+    }
     @IBAction func btnNext(_ sender: Any) {
-        switch backupFlowType {
+        switch viewModel.flowType {
         case .phrase:
             if let vc = AccountNavigator.mnemonic() {
                 navigationController?.pushViewController(vc, animated: true)
             }
         case .quiz:
-            if let vc = AccountNavigator.recover() {
-                navigationController?.pushViewController(vc, animated: true)
+            Task {
+                if let credentials = try? await wm.prominentSession?.getCredentials(password: "") {
+                    if let mnemonic = credentials.mnemonic {
+                        if let vc = AccountNavigator.phraseNoteDown(phrase: mnemonic) {
+                            navigationController?.pushViewController(vc, animated: true)
+                        }
+                    }
+                }
             }
+        case .addSubaccount:
+            selectLength()
+        }
+    }
+}
+extension ManualBackupViewController: DialogListViewControllerDelegate {
+    func didSwitchAtIndex(index: Int, isOn: Bool, type: DialogType) {}
+
+    func didSelectIndex(_ index: Int, with type: DialogType) {
+        switch PhrasePrefs(rawValue: index) {
+        case ._12:
+            next(._12)
+        case ._24:
+            next(._24)
+        case .none:
+            break
         }
     }
 }