+++ gaios/WalletTabBar/Tabs/Settings/TabSettingsModel.swift
@@ -158,7 +158,7 @@ class TabSettingsModel {
             menu += [twoFactorAuth, pgpKey]
         }
         if !isWatchonly {
-            menu += [watchOnly, archievedAccounts]
+            menu += [watchOnly, archievedAccounts, createAccount]
         }
         return menu
     }